+++ rl_blox/algorithm/sac.py
@@ -1,6 +1,5 @@
 from collections import namedtuple
 
-import chex
 import distrax
 import gymnasium as gym
 import jax
@@ -13,76 +12,7 @@ from flax import nnx
 from ..logging.logger import LoggerBase
 from .ddpg import ReplayBuffer, mse_action_value_loss, update_target
 from ..blox.function_approximator.mlp import MLP
-
-
-# TODO consolidate implementations
-class GaussianMLP(nnx.Module):
-    """Probabilistic neural network that predicts a Gaussian distribution.
-
-    Parameters
-    ----------
-    shared_head
-        All nodes of the last hidden layer are connected to mean AND log_std.
-
-    n_features
-        Number of features.
-
-    n_outputs
-        Number of output components.
-
-    hidden_nodes
-        Numbers of hidden nodes of the MLP.
-
-    rngs
-        Random number generator.
-    """
-
-    shared_head: bool
-    n_outputs: int
-    hidden_layers: list[nnx.Linear]
-    output_layers: list[nnx.Linear]
-
-    def __init__(
-        self,
-        shared_head: bool,
-        n_features: int,
-        n_outputs: int,
-        hidden_nodes: list[int],
-        rngs: nnx.Rngs,
-    ):
-        chex.assert_scalar_positive(n_features)
-        chex.assert_scalar_positive(n_outputs)
-
-        self.shared_head = shared_head
-        self.n_outputs = n_outputs
-
-        self.hidden_layers = []
-        n_in = n_features
-        for n_out in hidden_nodes:
-            self.hidden_layers.append(nnx.Linear(n_in, n_out, rngs=rngs))
-            n_in = n_out
-
-        self.output_layers = []
-        if shared_head:
-            self.output_layers.append(
-                nnx.Linear(n_in, 2 * n_outputs, rngs=rngs)
-            )
-        else:
-            self.output_layers.append(nnx.Linear(n_in, n_outputs, rngs=rngs))
-            self.output_layers.append(nnx.Linear(n_in, n_outputs, rngs=rngs))
-
-    def __call__(self, x: jnp.ndarray) -> tuple[jnp.ndarray, jnp.ndarray]:
-        for layer in self.hidden_layers:
-            x = nnx.swish(layer(x))
-
-        if self.shared_head:
-            y = self.output_layers[0](x)
-            mean, log_var = jnp.split(y, (self.n_outputs,), axis=-1)
-        else:
-            mean = self.output_layers[0](x)
-            log_var = self.output_layers[1](x)
-
-        return mean, log_var
+from ..blox.function_approximator.gaussian_mlp import GaussianMLP
 
 
 # TODO merge with Gaussian policy from REINFORCE branch
@@ -337,6 +267,7 @@ def create_sac_state(
     env: gym.Env[gym.spaces.Box, gym.spaces.Box],
     policy_shared_head: bool = False,
     policy_hidden_nodes: list[int] | tuple[int] = (256, 256),
+    policy_activation: str = "swish",
     policy_learning_rate: float = 3e-4,
     q_hidden_nodes: list[int] | tuple[int] = (256, 256),
     q_activation: str = "relu",
@@ -351,6 +282,7 @@ def create_sac_state(
         env.observation_space.shape[0],
         env.action_space.shape[0],
         policy_hidden_nodes,
+        policy_activation,
         nnx.Rngs(seed),
     )
     policy = GaussianPolicy(policy_net, env.action_space)
