+++ rl_blox/blox/function_approximator/gaussian_mlp.py
@@ -5,18 +5,21 @@ import jax.numpy as jnp
 from flax import nnx
 
 
-class MLP(nnx.Module):
-    """Multilayer Perceptron.
+class GaussianMLP(nnx.Module):
+    """Probabilistic neural network that predicts a Gaussian distribution.
 
     Parameters
     ----------
+    shared_head : bool
+        All nodes of the last hidden layer are connected to mean AND log_std.
+
     n_features : int
         Number of features.
 
     n_outputs : int
         Number of output components.
 
-    hidden_nodes : list
+    hidden_nodes : list[int]
         Numbers of hidden nodes of the MLP.
 
     activation : str
@@ -28,6 +31,9 @@ class MLP(nnx.Module):
 
     Attributes
     ----------
+    shared_head : bool
+        All nodes of the last hidden layer are connected to mean AND log_std.
+
     n_outputs : int
         Number of output components.
 
@@ -37,17 +43,19 @@ class MLP(nnx.Module):
     hidden_layers : list[nnx.Linear]
         Hidden layers.
 
-    output_layer : nnx.Linear
-        Output layer.
+    output_layers : list[nnx.Linear]
+        Output layers.
     """
 
+    shared_head: bool
     n_outputs: int
     activation: Callable[[jnp.ndarray], jnp.ndarray]
     hidden_layers: list[nnx.Linear]
-    output_layer: nnx.Linear
+    output_layers: list[nnx.Linear]
 
     def __init__(
         self,
+        shared_head: bool,
         n_features: int,
         n_outputs: int,
         hidden_nodes: list[int],
@@ -57,6 +65,7 @@ class MLP(nnx.Module):
         chex.assert_scalar_positive(n_features)
         chex.assert_scalar_positive(n_outputs)
 
+        self.shared_head = shared_head
         self.n_outputs = n_outputs
         self.activation = getattr(nnx, activation)
 
@@ -66,9 +75,24 @@ class MLP(nnx.Module):
             self.hidden_layers.append(nnx.Linear(n_in, n_out, rngs=rngs))
             n_in = n_out
 
-        self.output_layer = nnx.Linear(n_in, n_outputs, rngs=rngs)
+        self.output_layers = []
+        if shared_head:
+            self.output_layers.append(
+                nnx.Linear(n_in, 2 * n_outputs, rngs=rngs)
+            )
+        else:
+            self.output_layers.append(nnx.Linear(n_in, n_outputs, rngs=rngs))
+            self.output_layers.append(nnx.Linear(n_in, n_outputs, rngs=rngs))
 
-    def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
+    def __call__(self, x: jnp.ndarray) -> tuple[jnp.ndarray, jnp.ndarray]:
         for layer in self.hidden_layers:
             x = self.activation(layer(x))
-        return self.output_layer(x)
+
+        if self.shared_head:
+            y = self.output_layers[0](x)
+            mean, log_var = jnp.split(y, (self.n_outputs,), axis=-1)
+        else:
+            mean = self.output_layers[0](x)
+            log_var = self.output_layers[1](x)
+
+        return mean, log_var
